Hot Reload runs on client env
