+++ server/config/middleware.js
@@ -3,41 +3,13 @@
 const bodyParser = require('body-parser');
 const morgan = require('morgan');
 const passport = require('./passport');
-
-// create webpack middleware
-const webpack = require('webpack');
-const webpackconfig = require('../../webpack.config.js');
-
-// hooks webpack up to express
-const webpackDevMiddleware = require('webpack-dev-middleware');
-// hooks webpack up to hot module reloading
-const webpackHotMiddleware = require('webpack-hot-middleware');
-const compiler = webpack(webpackconfig);
-
-//hot module reloading middleware
-var useWebpackMiddleware = function (app) {
-  app.use(webpackDevMiddleware(compiler, {
-    publicPath: webpackconfig.output.publicPath,
-    noInfo: true,
-    stats: {
-      colors: true,
-      chunks: false,
-      'errors-only': true
-    }
-  }));
-  app.use(webpackHotMiddleware(compiler, {
-    log: console.log
-  }));
-
-  return app;
-};
+const hotreload = require('./hotreload');
 
 module.exports = (app, express) => {
 //  app.use(morgan('dev'));
-
-  useWebpackMiddleware(app);
+  process.env.NODE_ENV === 'client' && hotreload(app);
+  process.env.NODE_ENV === 'dev' && console.log('NOT running hotreload');
   app.use(passport.initialize());
   app.use(bodyParser.json());
   app.use(express.static(__dirname + '/../../client'));
-
 };
